# swagger.yaml
swagger: '2.0'
info:
  description: 'discretionary-business-grants API Schema'
  version: '0.0.31'
  title: 'discretionary-business-grants API'
host: 'discretionarybusinessgrants.hackney.gov.uk'
basePath: '/api'
...
schemes:
  - 'https'
paths:
...
  /applications:
    post:
      tags:
        - 'applications'
      summary: 'Create a new business grant application'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - name: 'body'
          in: 'body'
          required: true
          schema:
            $ref: '#/definitions/Application'
      responses:
        '201':
          description: 'Your application has been submitted'
          schema:
            $ref: '#/definitions/ApplicationsResponse'
        '400':
          description: 'Invalid input'
        '500':
          description: 'Internal server error'
    get:
      tags:
        - 'applications'
      summary: 'Get a list of applications'
      produces:
        - 'application/json'
      parameters:
        - in: 'query'
          name: 'page'
          type: 'number'
          required: false
          default: 1
        - in: 'query'
          name: 'pageSize'
          type: 'number'
          required: false
          default: 10
        - in: 'query'
          name: 'sort'
          type: 'string'
          required: false
          enum:
            [+applicationDate, -applicationDate, +businessName, -businessName]
          default: '+applicationDate'
        - in: 'query'
          name: 'status'
          type: 'string'
          required: false
        - in: 'query'
          name: 'businessType'
          type: 'string'
          required: false
        - in: 'query'
          name: 'applicationId'
          type: 'string'
          required: false
      responses:
        '200':
          description: 'List of applications'
          schema:
            $ref: '#/definitions/ApplicationsListResponse'
        '400':
          description: 'Invalid input'
        '500':
          description: 'Internal server error'

    patch:
      tags:
        - 'applications'
      summary: 'Export applications for payment. Produces csv report'
      produces:
        - 'text/csv'
      responses:
        '200':
          description: 'CSV report file of exported applications'
          schema:
            type: file
        '500':
          description: 'Internal server error'
...
definitions:
...
  Application:
    type: 'object'
    properties:
      clientGeneratedId:
        type: 'string'
        minLength: 1
      eligibilityCriteria:
        $ref: '#/definitions/EligibilityCriteria'
      contact:
        $ref: '#/definitions/Contact'
      business:
        $ref: '#/definitions/Business'
      turnover:
        $ref: '#/definitions/Turnover'
      fixedPropertyCosts:
        $ref: '#/definitions/FixedPropertyCosts'
      documents:
        $ref: '#/definitions/Documents'
      businessBankAccount:
        $ref: '#/definitions/BusinessBankAccount'
      declaration:
        $ref: '#/definitions/Declaration'
...
  EligibilityCriteria:
    type: 'object'
    properties:
      tradingInHackney:
        type: 'string'
      businessSizeId:
        description: 'The business must be classed as 1)Micro or 2)Small'
        type: 'string'
        minimum: 1
      typeOfBusinessId:
        type: 'string'
      tradingOn20200311:
        type: 'string'
      servedLegalNotices:
        type: 'string'
      receivedOtherGrants:
        type: 'string'
      hasFixedPropertyCost:
        type: 'string'
      significantIncomeFall:
        type: 'string'
    required:
      - tradingInHackney
      - typeOfBusinessId
      - businessSizeId
      - tradingOn20200311
      - servedLegalNotices
      - receivedOtherGrants
      - hasFixedPropertyCost
      - significantIncomeFall
  Contact:
    type: 'object'
    properties:
      contactTypeId:
        type: 'string'
      firstName:
        type: 'string'
        minLength: 1
      lastName:
        type: 'string'
        minLength: 1
      emailAddress:
        type: 'string'
        minLength: 5
      telephoneNumber:
        type: 'string'
        minLength: 5
      address:
        $ref: '#/definitions/ContactAddress'
    required:
      - contactTypeId
      - firstName
      - lastName
      - emailAddress
      - address
  ContactAddress:
    type: 'object'
    properties:
      uprn:
        type: 'string'
      firstLine:
        type: 'string'
        minLength: 1
      secondLine:
        type: 'string'
      thirdLine:
        type: 'string'
      postcode:
        type: 'string'
      ward:
        type: 'string'
    required:
      - firstLine
  Business:
    type: 'object'
    properties:
      companyStructureId:
        type: 'string'
      registeredName:
        description: 'Registered name at companies house'
        type: 'string'
      businessName:
        description: 'EG: Trading name, Charity name'
        type: 'string'
      businessDescription:
        description: 'What is the main activity carried out'
        type: 'string'
      companyNumber:
        type: 'string'
      siteDescriptionId:
        type: 'string'
      fullTimeEmployees:
        type: 'integer'
        minimum: 1
      percentageFallInIncome:
        minimum: 1
        maximum: 100
      rateableValue:
        type: 'number'
      ratesAccountNumber:
        type: 'string'
        description: 'Business Rates Account number'
      registeredCharity:
        type: 'string'
      councilRentAccountNumber:
        type: 'string'
      councilTaxNumber:
        type: 'string'
      businessAddress:
        $ref: '#/definitions/BusinessAddress'
    required:
      - companyStructureId
      - siteDescriptionId
      - fullTimeEmployees
      - percentageFallInIncome
      - businessAddress
  BusinessAddress:
    type: 'object'
    properties:
      uprn:
        type: 'string'
        minLength: 1
      firstLine:
        type: 'string'
        minLength: 1
      secondLine:
        type: 'string'
      thirdLine:
        type: 'string'
      postcode:
        type: 'string'
        minLength: 2
    required:
      - uprn
      - firstLine
      - postcode
  Turnover:
    type: 'object'
    properties:
      turnover:
        description: 'Business turnover March to May (inclusive) 2020'
        type: 'number'
      year1819:
        type: 'number'
      year1920:
        type: 'number'
    required:
      - year1819
      - year1920
  FixedPropertyCosts:
    type: 'object'
    properties:
      year2018To2019:
        type: 'number'
      year2019To2020:
        type: 'number'
      itemsIncluded:
        type: 'string'
    required:
      - year2018To2019
      - year2019To2020
      - itemsIncluded
  Documents:
    type: 'array'
    items:
      type: 'object'
      uniqueItems: true
      properties:
        s3Path:
          type: 'string'
          minLength: 1
        documentType:
          type: 'string'
          minLength: 1
      required:
        - s3Path
        - documentType
  BusinessBankAccount:
    type: 'object'
    properties:
      bankName:
        type: 'string'
        minLength: 1
      accountHolder:
        type: 'string'
        minLength: 1
      accountNumber:
        type: 'string'
        minLength: 1
      accountSortcode:
        type: 'string'
        minLength: 6
    required:
      - bankName
      - accountHolder
      - accountNumber
      - accountSortcode
  Declaration:
    type: 'object'
    properties:
      stateAidOptionId:
        type: 'string'
      dateOfAid:
        type: 'string'
      organisationProvidingAid:
        type: 'string'
      stateAidReceived:
        type: 'number'
      permittedToAcceptStateAidGrant:
        type: 'string'
      readUnderstoodDeclaration:
        type: 'boolean'
    required:
      - stateAidOptionId
      - readUnderstoodDeclaration
  ApplicationsResponse:
    type: 'object'
    properties:
      confirmationId:
        type: 'string'
  ApplicationsListResponse:
    type: 'object'
    properties:
      applications:
        type: 'array'
        items:
          type: 'object'
          uniqueItems: true
          properties:
            clientGeneratedId:
              type: 'string'
            businessName:
              type: 'string'
            applicationDate:
              type: 'string'
              format: 'date-time'
            status:
              type: 'string'
      pagination:
        $ref: '#/definitions/PaginationResponse'
      error:
        type: 'string'
...
  PaginationResponse:
    type: object
    properties:
      totalPages:
        type: number
      currentPage:
        type: number
      links:
        type: object
        properties:
          firstPage:
            type: string
          lastPage:
            type: string
          previousPage:
            type: string
          nextPage:
            type: string
...